refactor: split messages.py from init.py
